# contracts/openapi/openapi.yaml
openapi: 3.0.3
info:
  title: 知序云 API
  version: 0.1.0
servers:
  - url: /api/v1
paths:
...
  /admin/dashboard:
    get:
      operationId: getAdminDashboard
      security: [{ bearerAuth: [] }]
      responses:
        '200': { description: 管理员首页统计与系统状态 }
        '403': { description: 需要管理员角色 }
  /admin/users:
    get:
      operationId: searchAdminUsers
      security: [{ bearerAuth: [] }]
      parameters:
        - { name: page, in: query, schema: { type: integer, default: 0 } }
        - { name: size, in: query, schema: { type: integer, default: 10 } }
        - { name: keyword, in: query, schema: { type: string } }
        - { name: role, in: query, schema: { type: string, enum: [STUDENT, TEACHER, ADMIN] } }
      responses:
        '200': { description: 管理员账号分页查询 }
    post:
      operationId: createAdminUser
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content: { application/json: { schema: { type: object, additionalProperties: true } } }
      responses:
        '200': { description: 创建账号 }
  /admin/users/{userId}:
    patch:
      operationId: updateAdminUser
      security: [{ bearerAuth: [] }]
      parameters: [{ name: userId, in: path, required: true, schema: { type: integer, format: int64 } }]
      requestBody:
        required: true
        content: { application/json: { schema: { type: object, additionalProperties: true } } }
      responses:
        '200': { description: 更新账号资料 }
  /admin/users/{userId}/status:
    patch:
      operationId: setAdminUserStatus
      security: [{ bearerAuth: [] }]
      parameters: [{ name: userId, in: path, required: true, schema: { type: integer, format: int64 } }]
      responses:
        '200': { description: 更新账号启用状态 }
  /admin/users/{userId}/password:
    patch:
      operationId: resetAdminUserPassword
      security: [{ bearerAuth: [] }]
      parameters: [{ name: userId, in: path, required: true, schema: { type: integer, format: int64 } }]
      responses:
        '204': { description: 重置账号密码 }
  /admin/courses:
    get:
      operationId: listAdminCourses
      security: [{ bearerAuth: [] }]
      responses:
        '200': { description: 管理员课程目录 }
    post:
      operationId: createAdminCourse
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content: { application/json: { schema: { type: object, additionalProperties: true } } }
      responses:
        '200': { description: 创建课程 }
  /admin/courses/{courseId}/classes:
    get:
      operationId: listAdminCourseClasses
      security: [{ bearerAuth: [] }]
      parameters: [{ name: courseId, in: path, required: true, schema: { type: integer, format: int64 } }]
      responses:
        '200': { description: 查询课程教学班 }
    post:
      operationId: createAdminCourseClass
      security: [{ bearerAuth: [] }]
      parameters: [{ name: courseId, in: path, required: true, schema: { type: integer, format: int64 } }]
      requestBody:
        required: true
        content: { application/json: { schema: { type: object, additionalProperties: true } } }
      responses:
        '200': { description: 创建教学班 }
  /admin/administrative-classes:
    get:
      operationId: listAdministrativeClasses
      security: [{ bearerAuth: [] }]
      responses:
        '200': { description: 行政班目录 }
    post:
      operationId: createAdministrativeClass
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content: { application/json: { schema: { type: object, additionalProperties: true } } }
      responses:
        '200': { description: 创建行政班 }
  /admin/settings:
    get:
      operationId: getAdminSettings
      security: [{ bearerAuth: [] }]
      responses:
        '200': { description: 管理员可见平台设置 }
    put:
      operationId: updateAdminSettings
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content: { application/json: { schema: { type: object, additionalProperties: true } } }
      responses:
        '200': { description: 保存平台设置 }
  /admin/announcements:
    post:
      operationId: publishAdminAnnouncement
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [scope, level, title, content]
              properties:
                scope: { type: string, enum: [全部用户, 仅教师, 仅学生] }
                level: { type: string }
                title: { type: string, maxLength: 220 }
                content: { type: string, maxLength: 1000 }
      responses:
        '200': { description: 消息已投递给目标用户 }
...
  /admin/logs:
    get:
      operationId: listAdminAuditLogs
      security: [{ bearerAuth: [] }]
      parameters:
        - { name: page, in: query, schema: { type: integer, default: 0 } }
        - { name: size, in: query, schema: { type: integer, default: 20 } }
        - { name: keyword, in: query, schema: { type: string } }
      responses:
        '200': { description: 管理操作日志分页结果 }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
...